Updated the appName component to include menuAction for the html app.
Updated the FileAction component to include the kind property in the payload send to the closeFile action creator.

Added the FrameAction component which will handle the display of the Safari appName and the diplsya of Safair menuAction

Updated the componentDidMount of the Safari component to include the retrieve the kind property from the props. This was done to add it in the payload send by the closeFrame